+++ CHANGELOG.md
@@ -1,6 +1,14 @@
 # Changelog
 
-- Improved compile error if the `texture_atlas` or `standard_material` attributes are used without the render feature
+- Split `render` feature into `2d` (for texture atlas support) and `3d` (standard material support)
+- Support loading folders as vectors of typed handles ([#30](https://github.com/NiklasEi/bevy_asset_loader/issues/30))
+- 'Folder' can be used as alias for 'File' in dynamic asset collection ron files
+
+## v0.9.0
+- Folder assets are now marked with the `folder` attribute. The path is defined as for any other asset in `path`.
+  - This adds support for dynamic folders in asset collections ([#28](https://github.com/NiklasEi/bevy_asset_loader/issues/28))
+- Support optional dynamic assets in collections ([#24](https://github.com/NiklasEi/bevy_asset_loader/issues/24))
+- Improved compile error if the `texture_atlas` or `standard_material` attributes are used without the render feature (related to [#27](https://github.com/NiklasEi/bevy_asset_loader/issues/27))
 - Support loading keys for dynamic assets from ron files
   - New example `dynamic_asset_ron`
 - Support initialising asset collections directly on the bevy App or World